Define locations for config files and data dir.
This documents the search order for the configuration file and data
directory. It defines new support for the XDG Base Directory
Specification and additional options for overriding the data directory
on the command line or in the configuration file. For symmetry, it also
adds an alias `spooldir` for `spool` in the configuration file.